Upgrade to Java 21 (#7294)
* Bump Java version from 17 to 21
* Change `@Retryable` to not use deprecated `value` attribute
* Fix compiler warnings
* Remove unused Query DSL dependency
* Suppress compiler warnings due to plugin generated code

---------

+++ hedera-mirror-web3/src/main/java/com/hedera/services/fees/usage/token/TokenOpsUsageUtils.java
@@ -190,7 +190,7 @@ public enum TokenOpsUsageUtils {
         return baseSize;
     }
 
-    public static <T> long keySizeIfPresent(final T op, final Predicate<T> check, final Function<T, Key> getter) {
-        return check.test(op) ? getAccountKeyStorageSize(getter.apply(op)) : 0L;
+    public static <T> int keySizeIfPresent(final T op, final Predicate<T> check, final Function<T, Key> getter) {
+        return check.test(op) ? getAccountKeyStorageSize(getter.apply(op)) : 0;
     }
 }
